EncodeToMorseCode brief explanation and encodeWord fix in encoder/encoder.go
Added explanation of how it works, and what it does to the EncodeToMorseCode
to be seen when the developers' cursor stands on the function. encodeWord
function were always adding one more letter splitter character to the end.
This issue fixed by removing last element of the returning encoded word
after loop and before return.