TestIGetAStringAsParameterSuccessfullyFromRequest test is created and completed
In the test, I sent a test text 'AB' to the fiber app (which I writed a handler
for post operation with query param 'text') and catched it in the post method
and stored in a test-scoped global string variable. Than compared test text with
this string variable to be sure that the app is taking the text parameter from
query correctly.