TestISentGivenStringInMorseEncodedFormatSuccessfully test is created and tested.
In the test, I sent a test string 'AB' to the created api which has a handler
for a get request with query params. This handler takes this query params and
returns its morse encoded format in response. I tested the test string's encoded
format exactly same with the returned encoded string in response of test request's
body. Test is passed.